% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longstring.R
\name{ALSP}
\alias{ALSP}
\title{calculate adaptive longstrongpattern}
\usage{
ALSP(x, maxlen = 5)
}
\arguments{
\item{x}{a vector of responses of a given respondent}

\item{maxlen}{maximum length that a pattern can encompass}
}
\description{
calculate adaptive longstrongpattern
}
