% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencoder}
\alias{autoencoder}
\title{fit an autoencoder}
\usage{
autoencoder(
  data,
  hidden_layers = c(10, 2, 10),
  activation = c("tanh", "linear", "tanh"),
  loss = get_pseudo_huber(),
  optimizer = keras::optimizer_sgd(learning_rate = 1e-04),
  kernel_regularizer_HL1 = NULL,
  bias_regularizer_HL1 = NULL,
  epochs = 100L,
  batch_size = 10L,
  verbose = 2L,
  seed = NULL
)
}
\arguments{
\item{data}{data matrix that holds the responses of a given respondent in its rows}

\item{hidden_layers}{specify number of hidden layers and the number of nodes therein}

\item{activation}{specify activation function in each hidden layer}

\item{loss}{a function object for the loss function}

\item{optimizer}{keras object for the optimizer}

\item{kernel_regularizer_HL1}{regularization imposed in hidden layer on weights}

\item{bias_regularizer_HL1}{regularization imposed in hidden layer on biases}

\item{epochs}{number of epochs}

\item{batch_size}{batch size}

\item{verbose}{manage prints}

\item{seed}{random seed}
}
\description{
fit an autoencoder
}
