% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{changepoint_multivariate}
\alias{changepoint_multivariate}
\title{Detects at most one changepoint in a multidimensional time series}
\usage{
changepoint_multivariate(x, Kn, theta = "mean")
}
\arguments{
\item{x}{A matrix of size (d x n)}

\item{Kn}{a vector of critical values}

\item{theta}{statistic along which to look for changepoints. Either "mean" or "median"}
}
\description{
The matrix x holds n time series observations (each of dimension d) in its columns, so x is (d x n)
}
