% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{changepoint_univariate}
\alias{changepoint_univariate}
\title{Detects at most one changepoint in a univariate time series}
\usage{
changepoint_univariate(x, Kn = 40.1, theta = "mean")
}
\arguments{
\item{x}{A vector of time series observations}

\item{Kn}{a vector of critical values}

\item{theta}{statistic along which to look for changepoints. Either "mean" or "median"}
}
\description{
Detects at most one changepoint in a univariate time series
}
