% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints.R
\name{detect_changepoints_multivariate}
\alias{detect_changepoints_multivariate}
\title{#' Detect changepoints across multiple multivariate series}
\usage{
detect_changepoints_multivariate(
  data,
  alpha = c(0.005, 0.001),
  theta = "mean",
  mc_cores = parallel::detectCores(),
  matrix = FALSE,
  teststat = FALSE
)
}
\arguments{
\item{data}{A list that contains matrices, each of which holds n time series observations (each of dimension d) in its columns, so each matrix is (d x n). We search for changepoints along these data}

\item{alpha}{vector of significance levels}

\item{theta}{statistic along which to look for changepoints. Either "mean" or "median"}

\item{mc_cores}{number of cores to run in parallel (=1 for no parallelization). Note: Parallelization is currently only supported for Unix-like systems}

\item{matrix}{logical. If TRUE, then a binary matrix of the dimensionality of the input data is returned in which each 1 represents a changepoint}

\item{teststat}{Logical. Shall the test statistics for each period be returned?}
}
\description{
#' Detect changepoints across multiple multivariate series
}
