% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints.R
\name{detect_changepoints_univariate}
\alias{detect_changepoints_univariate}
\title{Detect changepoints across multiple univariate series}
\usage{
detect_changepoints_univariate(
  data,
  alpha = c(0.005, 0.001),
  theta = "mean",
  mc_cores = parallel::detectCores(),
  matrix = FALSE,
  teststat = FALSE,
  CP_at_segment_end = TRUE
)
}
\arguments{
\item{data}{data matrix, each row of which represents a time series along which a changepoint is searched}

\item{alpha}{vector of significance levels}

\item{theta}{statistic along which to look for changepoints. Either "mean" or "median"}

\item{mc_cores}{number of cores to run in parallel (=1 for no parallelization). Note: Parallelization is currently only supported for Unix-like systems}

\item{matrix}{logical. If TRUE, then a binary matrix of the dimensionality of the input data is returned in which each 1 represents a changepoint}

\item{teststat}{Logical. Shall the test statistics for each period be returned?}

\item{CP_at_segment_end}{Logical. Should the changepoint be the last period of an ending segment (TRUE, default, as in Shao et al.) or the first period of a beginning segment (FALSE)?}
}
\description{
Detect changepoints across multiple univariate series
}
