% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_onset}
\alias{get_onset}
\title{retrieve estimated careless onset items}
\usage{
get_onset(x, alpha = 0.001)
}
\arguments{
\item{x}{object of class \code{carelessonset}}

\item{alpha}{significance levels}
}
\description{
retrieve estimated careless onset items
}
