% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.carelessonset}
\alias{plot.carelessonset}
\title{plot method for a given observation}
\usage{
\method{plot}{carelessonset}(x, idx = 1, alpha = 0.001, ...)
}
\arguments{
\item{x}{object of class \code{carelessonset}}

\item{idx}{index}

\item{alpha}{significance levels}

\item{...}{additional parameters to be passed down}
}
\description{
plot method for a given observation
}
