% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dimension}
\alias{plot_dimension}
\title{visualize dimensions}
\usage{
plot_dimension(
  RE = NULL,
  ALSP = NULL,
  time = NULL,
  onset = NULL,
  color = "blue"
)
}
\arguments{
\item{RE}{vector of reconstruction errors}

\item{ALSP}{vector of ALSP values}

\item{time}{vector of response times}

\item{onset}{index of carelessness onset}

\item{color}{color of vertical line denoting onset}
}
\description{
visualize dimensions
}
