% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_teststat}
\alias{plot_teststat}
\title{visualize SNCP test statistic over all periods}
\usage{
plot_teststat(x, d, alpha = c(0.05, 0.025, 0.01))
}
\arguments{
\item{x}{a vector of test statistics of length \code{p-1}}

\item{d}{dimension of the initial series}

\item{alpha}{vector of significance levels; possibly \code{NULL}}
}
\description{
visualize SNCP test statistic over all periods
}
